project 2 3rd attempt.. get the histogram priv to be accurate

+++ SDH.cu
@@ -360,7 +360,7 @@ __global__ void PDH_kernel4(unsigned long long* d_histogram,
 					Rz = Lz - Rz;
 					Rz *= Rz;
 
-					dist = sqrt((Rx) + (Ry) + (Rz));
+					dist = sqrtf((Rx) + (Ry) + (Rz));
 					h_pos = (int)(dist/res);
 					/* END DISTANCE FUNCTION */
 
@@ -398,7 +398,7 @@ __global__ void PDH_kernel4(unsigned long long* d_histogram,
 				Rz = Lz - Rz;
 				Rz *= Rz;
 				
-				dist = sqrt((Rx) + (Ry) + (Rz));
+				dist = sqrtf((Rx) + (Ry) + (Rz));
 				/* END DISTANCE FUNCTION */
 
 				h_pos = (int)(dist/res);
